refactor: Improve project and class orginization
- Renamed OneNoteQuickFormatter to NoteQuickFormatter.
- Removed Windows Forms App project.
- Renamed classes for better clarity.
