+++ NoteQuickFormatter.sln
@@ -3,13 +3,11 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 17
 VisualStudioVersion = 17.13.35825.156
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "NoteQuickFormatter", "NoteQuickFormatter\NoteQuickFormatter.csproj", "{D5FE3003-9B11-4F6B-BACF-33842E0CECF3}"
-EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "NoteQuickFormatterTests", "NoteQuickFormatterTests\NoteQuickFormatterTests.csproj", "{0707EAE9-8BEE-4D6D-B8D6-1AA8982A0AC5}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "OneNoteQuickFormatter", "OneNoteQuickFormatter\OneNoteQuickFormatter.csproj", "{DFA8B735-4028-4953-BB78-671C2DD9B4FD}"
+Project("{54435603-DBB4-11D2-8724-00A0C9A8B90C}") = "Setup", "Setup\Setup.vdproj", "{F5D91FD6-F519-3AC5-2EF1-6ADD62B256F6}"
 EndProject
-Project("{54435603-DBB4-11D2-8724-00A0C9A8B90C}") = "Setup1", "Setup1\Setup1.vdproj", "{F5D91FD6-F519-3AC5-2EF1-6ADD62B256F6}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "NoteQuickFormatter", "NoteQuickFormatter\NoteQuickFormatter.csproj", "{DFA8B735-4028-4953-BB78-671C2DD9B4FD}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -17,20 +15,16 @@ Global
 		Release|Any CPU = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{D5FE3003-9B11-4F6B-BACF-33842E0CECF3}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{D5FE3003-9B11-4F6B-BACF-33842E0CECF3}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{D5FE3003-9B11-4F6B-BACF-33842E0CECF3}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{D5FE3003-9B11-4F6B-BACF-33842E0CECF3}.Release|Any CPU.Build.0 = Release|Any CPU
 		{0707EAE9-8BEE-4D6D-B8D6-1AA8982A0AC5}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{0707EAE9-8BEE-4D6D-B8D6-1AA8982A0AC5}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{0707EAE9-8BEE-4D6D-B8D6-1AA8982A0AC5}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{0707EAE9-8BEE-4D6D-B8D6-1AA8982A0AC5}.Release|Any CPU.Build.0 = Release|Any CPU
+		{F5D91FD6-F519-3AC5-2EF1-6ADD62B256F6}.Debug|Any CPU.ActiveCfg = Debug
+		{F5D91FD6-F519-3AC5-2EF1-6ADD62B256F6}.Release|Any CPU.ActiveCfg = Release
 		{DFA8B735-4028-4953-BB78-671C2DD9B4FD}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{DFA8B735-4028-4953-BB78-671C2DD9B4FD}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{DFA8B735-4028-4953-BB78-671C2DD9B4FD}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{DFA8B735-4028-4953-BB78-671C2DD9B4FD}.Release|Any CPU.Build.0 = Release|Any CPU
-		{F5D91FD6-F519-3AC5-2EF1-6ADD62B256F6}.Debug|Any CPU.ActiveCfg = Debug
-		{F5D91FD6-F519-3AC5-2EF1-6ADD62B256F6}.Release|Any CPU.ActiveCfg = Release
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
