+++ NoteQuickFormatter/NextMonthToDoTemplateDialog.en.resx
