+++ NoteQuickFormatter/NextMonthToDoTemplateDialog.resx
