--- NoteQuickFormatter/Properties/Settings.settings
@@ -1,9 +1,0 @@
-﻿<?xml version='1.0' encoding='utf-8'?>
-<SettingsFile xmlns="http://schemas.microsoft.com/VisualStudio/2004/01/settings" CurrentProfile="(Default)" GeneratedClassNamespace="NoteQuickFormatter.Properties" GeneratedClassName="Settings">
-  <Profiles />
-  <Settings>
-    <Setting Name="Months" Type="System.String" Scope="User">
-      <Value Profile="(Default)">Jan;Feb;Mar;Apr;May;Jun;Jul;Aug;Sep;Oct;Nov;Dec</Value>
-    </Setting>
-  </Settings>
-</SettingsFile>
