+++ NoteQuickFormatter/ribbon.xml
